pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = ["Anton Manyanin <anton.manyanin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.65.2"}
databases = "^0.4.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.23.0"
psycopg2 = "^2.9.1"
alembic = "^1.6.5"
Random-Word = "^1.0.7"
requests = "^2.25.1"
requests-oauthlib = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
